Production-grade hardening pass — v1.0.0
Engineering-grade upgrade across every layer: backend resilience,
test coverage, ML evaluation tooling, frontend UX, and DevOps.
70 tests now pass (was 36), backend coverage 97 %, ruff lint clean.

Backend
* Request-ID middleware + X-Response-Time-ms; incoming X-Request-ID
  propagates for cross-service tracing
* Centralised ErrorResponse contract — every non-2xx is now typed JSON
* Structured per-request logging (request_id on every log line)
* Cache hygiene: prune_expired() on startup + inference_log GC (7 d)
  + cache_stats() surfaced via /api/health
* New /api/version with git revision + ML feature schema
* ML engine: defensive predict_rain_probability (NaN/Inf/None safe),
  improved heuristic uses pressure_change_3h
* Terrain: antimeridian wrap, polar clamp, ocean/no-data DEM cells
* Fire-and-forget cache.set/log_inference (response never waits on I/O)
* Caught tenacity.RetryError → clean 502 instead of 500

ML pipeline
* scripts/4_evaluate_model.py — publication-quality figures:
  ROC + AUC, PR + AP, calibration + Brier, threshold sweep,
  feature importance, confusion matrix at F2-optimal threshold
* figures/{evaluation_summary.json,threshold_sweep.csv} committed
* models/MODEL_CARD.md — HuggingFace-style model card with intended
  use, training data, evaluation (AUC 0.871 · AP 0.750 · Brier 0.138 ·
  F2 0.778 @ τ=0.20 · Recall 93.4 %), limitations, ethics
* Real ERA5 evaluation: 35 063 test samples, calibration adopted at
  τ=0.20 (F2-optimal) instead of default 0.5 — recall 93.4 %

Tests
* tests/test_api.py — HTTP integration with respx mocks: happy path,
  cache hit, per-activity cache slot, 422 invalid input, 502 upstream
  failure, CORS, OpenAPI schema
* tests/test_cache.py — TTL, expiry, prune, stats, per-activity keys
* tests/test_terrain_edge.py — antimeridian / poles / malformed DEM
* tests/test_ml_engine.py — fallback heuristic + NaN/None resilience
* tests/conftest.py — session-isolated MICROCLIMATEX_DB

Frontend
* Activity selector with aria-pressed + focus-visible; localStorage
* 4 mini-gauges with bilingual tooltips explaining what drives each
* R1-R4 decision-table badges (highlight on fire)
* Quick-scenarios dropdown: Genting · Cameron · Kinabalu · Everest · Singapore
* Loading spinner + slide-in toast (warns when ML model not loaded)
* Map layer switcher (Dark + Topographic)
* Mobile-responsive activity bar
* Bilingual EN/ZH UI persisted across reloads

DevOps
* .github/workflows/ci.yml: pytest matrix (3.9/3.11/3.12) + ruff +
  coverage XML artefact + Docker buildx build with GHA cache
* Multi-stage Dockerfile: builder → slim runtime as non-root mcx user
  with /api/health HEALTHCHECK
* docker-compose.yml with named data volume
* Makefile — make install-dev/test/lint/run/synth/preprocess/train/
  evaluate/docker/clean
* requirements-dev.txt — split dev tooling
* pyproject.toml — ruff config (E F I UP B SIM RUF) + pytest config
* .pre-commit-config.yaml — trailing-whitespace, EOF, ruff fix+format
* .dockerignore — keeps the image lean

Docs
* CHANGELOG.md (Keep a Changelog format) — 1.0.0 entry
* CONTRIBUTING.md — setup, PR checklist, safety-critical review notes
* README.md — CI badge, coverage badge, Makefile-driven quick start,
  Docker one-liner, updated training results pointing to MODEL_CARD.md

+++ backend/rule_engine.py
@@ -61,7 +61,7 @@ def _bin_level(score: int) -> RiskLevel:
 
 
 def _clip(x: float) -> int:
-    return max(0, min(100, int(round(x))))
+    return max(0, min(100, round(x)))
 
 
 # ════════════════════════════════════════════════════════════════════════
